The first working draft
In this, using Gemini, we first check if the doubt is related to our knowledge base(hard-coded questions and answers)  or if a web search is needed. This took some time, since the prompt had to be written in such a way as to achieve the desired functionality.

We can check the hard-coded questions by seeing that they work accordingly.

PS:- The key mentioned is the one I generated and has a limit. If it stops working, then just generate and attach your own key. Or contact me for any help.